dashboard/requirements.txt
fastapi
uvicorn[standard]
docker
jinja2
pydantic
python-dotenv
pytest
